Factor parent frame communication into own module.
