+++ app-src/index.js
@@ -1,7 +1,6 @@
 var initPlayerRouter = require('dce-player-router');
 var pathExists = require('object-path-exists');
-var createParentFrameListener = require('./create-parent-frame-listener');
-var createParentFrameAppriser = require('./create-parent-frame-appriser');
+var setUpParentFrameCommunications = require('./set-up-parent-frame-communications');
 
 var seekMethodPath = ['player', 'videoContainer', 'seekToTime'];
 
@@ -14,7 +13,6 @@ var router = initPlayerRouter({
 
 function seekWhenPaellaIsReady(startTime, endTime) {
   seek();
-  // TODO: Implement endTime support.
 
   function seek() {
     if (pathExists(paella, seekMethodPath)) {
@@ -44,27 +42,9 @@ function disableAutoHiding() {
 function noOp() {
 }
 
-function setUpParentFrameListener() {
-  $(document).off('paella:loadComplete', setUpParentFrameListener);
-  createParentFrameListener({
-    playResponder: playVideos
-  });
-}
-
-function setUpParentFrameAppriser() {
-  $(document).off('paella:loadComplete', setUpParentFrameAppriser);
-  createParentFrameAppriser({
-    videoElement: document.querySelector('#' + paella.player.videoContainer.video1Id)
-  });
-}
-
-function playVideos() {
-  paella.player.videoContainer.play();
-}
 
 ((function go() {
-  $(document).on('paella:loadComplete', setUpParentFrameListener);
-  $(document).on('paella:loadComplete', setUpParentFrameAppriser);
+  setUpParentFrameCommunications(document);
   disableAutoHiding();
   clearDoneUrlCookie();
   router.route();
